Rename Select to SelectWithHTMLOption
