+++ ui/src/ModalFormComponents/EditTagRow.tsx
@@ -25,7 +25,7 @@ import {TagRequest} from '../Tags/TagRequest.interface';
 
 import './TagRowsContainer.scss';
 import {Color} from '../Roles/RoleTag.interface';
-import {OptionType} from './Select';
+import {OptionType} from './SelectWithHTMLOptions';
 import ColorDropdown from '../Roles/ColorDropdown';
 import {RoleEditRequest} from '../Roles/RoleEditRequest.interface';
