+++ ui/src/ModalFormComponents/SelectWithHTMLOptions.scss
